// src/app/model/moives.model.ts
export namespace MoviesModel {
    
    export interface MoviesResponse {
        page: number
        results: Result[]
        total_pages: number
        total_results: number
    }
      
    export interface Result {
        adult: boolean
        backdrop_path: string
        genre_ids: number[]
        id: number
        original_language: string
        original_title: string
        overview: string
        popularity: number
        poster_path: string
        release_date: string
        title: string
        video: boolean
        vote_average: number
        vote_count: number
    }

    export interface MovieIdResponse {
        adult: boolean
        backdrop_path: string
        belongs_to_collection: any
        budget: number
        genres: Genre[]
        homepage: string
        id: number
        imdb_id: string
        original_language: string
        original_title: string
        overview: string
        popularity: number
        poster_path: string
        production_companies: ProductionCompany[]
        production_countries: ProductionCountry[]
        release_date: string
        revenue: number
        runtime: number
        spoken_languages: SpokenLanguage[]
        status: string
        tagline: string
        title: string
        video: boolean
        vote_average: number
        vote_count: number
    }
      
    interface Genre {
        id: number
        name: string
    }
      
    interface ProductionCompany {
        id: number
        logo_path?: string
        name: string
        origin_country: string
    }
      
    interface ProductionCountry {
        iso_3166_1: string
        name: string
    }
      
    interface SpokenLanguage {
        english_name: string
        iso_639_1: string
        name: string
    }
      
}  